divide action file to separate files in action folder
